# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BéninExpense AI - Plateforme de gestion des dépenses et conformité fiscale
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: getAccounts
      tags: [accounts]
      summary: List accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
...
  /expenses:
    get:
      operationId: getExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: accountId
          in: query
          schema:
            type: integer
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, validated, rejected, synced]
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"

    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get expense by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Not found

    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"

    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /momo/transactions:
    get:
      operationId: getMomoTransactions
      tags: [momo]
      summary: List MoMo transactions for the authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of MoMo transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MomoTransaction"
  /momo/simulate:
    post:
      operationId: simulateMomoTransaction
      tags: [momo]
      summary: Simulate an incoming MoMo transaction (demo)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MomoSimulateInput"
      responses:
        "201":
          description: Simulated transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MomoSimulateResult"
...
  /recurring:
    get:
      operationId: getRecurring
      tags: [recurring]
      summary: List recurring expenses
      responses:
        "200":
          description: List of recurring expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecurringExpense"
    post:
      operationId: createRecurring
      tags: [recurring]
      summary: Create a recurring expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurringExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurringExpense"
...
  /recurring/{id}:
    get:
      operationId: getRecurringById
      tags: [recurring]
      summary: Get recurring expense by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recurring expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurringExpense"
        "404":
          description: Not found
    put:
      operationId: updateRecurring
      tags: [recurring]
      summary: Update a recurring expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurringExpenseInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurringExpense"
    delete:
      operationId: deleteRecurring
      tags: [recurring]
      summary: Delete a recurring expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # DGI settings
...
components:
  schemas:
...
    Account:
      type: object
      required: [id, name, type, balance, currency]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [cash, bank, mobile_money]
        balance:
          type: number
        currency:
          type: string
          default: FCFA
        createdAt:
          type: string
          format: date-time
    AccountInput:
      type: object
      required: [name, type, balance]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [cash, bank, mobile_money]
        balance:
          type: number
        currency:
          type: string
          default: XOF
          description: ISO 4217 currency code (XOF, XAF, GHS, NGN, etc.)
...
    Expense:
      type: object
      required: [id, description, amount, category, status, createdAt]
      properties:
        id:
          type: integer
        description:
          type: string
        amount:
          type: number
        category:
          type: string
        status:
          type: string
          enum: [pending, validated, rejected, synced]
        accountId:
          type: ["integer", "null"]
        employeeId:
          type: ["integer", "null"]
        employeeName:
          type: ["string", "null"]
        receiptUrl:
          type: ["string", "null"]
        dgiQrCode:
          type: ["string", "null"]
        dgiReference:
          type: ["string", "null"]
        dgiStatus:
          type: string
          enum: [not_submitted, pending, normalized, failed]
        notes:
          type: ["string", "null"]
        submittedBy:
          type: ["integer", "null"]
          description: ID of the user who submitted the expense
        rejectionReason:
          type: ["string", "null"]
          description: Reason for rejection if status is rejected
        riskLevel:
          type: string
          enum: [none, low, medium, high]
          default: none
        flagReason:
          type: ["string", "null"]
          description: Reason for flagging the expense as suspicious
        createdAt:
          type: string
          format: date-time
    ExpenseInput:
      type: object
      required: [description, amount, category]
      properties:
        description:
          type: string
        amount:
          type: number
        category:
          type: string
        accountId:
          type: integer
        employeeId:
          type: integer
        notes:
          type: string
        receiptUrl:
          type: string
    ExpenseUpdate:
      type: object
      properties:
        description:
          type: string
        amount:
          type: number
        category:
          type: string
        status:
          type: string
          enum: [pending, validated, rejected]
        notes:
          type: string

    # Budget schemas
...
    MomoTransaction:
      type: object
      required: [id, userId, provider, providerLabel, phone, amount, type, description, reference, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        provider:
          type: string
          enum: [mtn_momo, moov_money, orange_money]
        providerLabel:
          type: string
        phone:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [debit, credit]
        description:
          type: string
        reference:
          type: string
        status:
          type: string
          enum: [pending, processed, ignored]
        expenseId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    MomoSimulateInput:
      type: object
      required: [amount, description]
      properties:
        provider:
          type: string
          enum: [mtn_momo, moov_money, orange_money]
          default: mtn_momo
        phone:
          type: string
          default: "229 97000000"
        amount:
          type: number
        type:
          type: string
          enum: [debit, credit]
          default: debit
        description:
          type: string
    MomoSimulateResult:
      type: object
      required: [status, transactionId, reference, providerLabel]
      properties:
        status:
          type: string
        transactionId:
          type: integer
        expenseId:
          type: ["integer", "null"]
        reference:
          type: string
        providerLabel:
          type: string

    # Recurring expense schemas
    RecurringExpense:
      type: object
      required: [id, description, amount, category, frequency, nextDueDate, active, createdAt]
      properties:
        id:
          type: integer
        description:
          type: string
        amount:
          type: number
        category:
          type: string
        frequency:
          type: string
          enum: [daily, weekly, monthly, yearly]
        nextDueDate:
          type: string
          format: date-time
        lastGeneratedAt:
          type: ["string", "null"]
          format: date-time
        active:
          type: boolean
        accountId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    RecurringExpenseInput:
      type: object
      required: [description, amount, category, frequency, nextDueDate]
      properties:
        description:
          type: string
        amount:
          type: number
        category:
          type: string
        frequency:
          type: string
          enum: [daily, weekly, monthly, yearly]
        nextDueDate:
          type: string
          format: date-time
        accountId:
          type: integer
        active:
          type: boolean
          default: true
        notes:
          type: string

    # DGI settings schemas
...